Update the "See New Tweets" to not return to home
Previously when the tweet section was set to a user's timline,
the "See New Tweets" click would send the tweet feed back to the
home feed with all the users tweets visible. This update will allow
the "See New Tweets" click to only show the user's new tweets when
the tweet feed is already set to that user's timeline.