Update the "Tweet!" clicker to not return home
Previously the visitor "Tweet!" clicker automatically returned to
the home feed with all the users tweets. With this commit, if the
tweet feed is set to the visitor's timeline, the feed will remain
on the visitor's timeline and only show the new tweet from the
visitor. If the feed is on a different user's timeline or the home
feed then the feed will show the home feed with the visitor's and
(other users') new tweet added to it.